+++ packages/graphiql-explorer/src/index.ts
@@ -1,2 +1,6 @@
 export type { Props as GraphiQLExplorerProps } from './explorer';
-export { Explorer } from './explorer';
+export {
+  Explorer as ExplorerInner,
+  ExplorerWrapper as default,
+  ExplorerWrapper as Explorer,
+} from './explorer';
