+++ packages/graphiql-plugin-explorer/package.json
@@ -25,11 +25,11 @@
   "scripts": {
     "prebuild": "rimraf dist types",
     "dev": "vite",
-    "build": "tsc --emitDeclarationOnly && node resources/copy-types.mjs && vite build && UMD=true vite build",
+    "build": "tsc --emitDeclarationOnly && vite build && UMD=true vite build",
     "preview": "vite preview"
   },
   "dependencies": {
-    "graphiql-explorer": "^0.9.0"
+    "graphiql-explorer": "^0.9.1"
   },
   "peerDependencies": {
     "@graphiql/react": "^0.20.1",
