+++ packages/graphiql-plugin-explorer/src/index.tsx
@@ -5,13 +5,14 @@ import {
   useSchemaContext,
   useOperationsEditorState,
 } from '@graphiql/react';
+
 import {
-  Explorer as GraphiQLExplorer,
-  GraphiQLExplorerProps,
+  ExplorerInner as GraphiQLExplorer,
+  type GraphiQLExplorerProps,
 } from 'graphiql-explorer';
+
 import React, { useCallback } from 'react';
 
-import './graphiql-explorer.d.ts';
 import './index.css';
 
 const colors = {
@@ -118,7 +119,7 @@ const styles = {
 
 export type GraphiQLExplorerPluginProps = Omit<
   GraphiQLExplorerProps,
-  'onEdit' | 'query'
+  'onEdit' | 'query' | 'explorerIsOpen'
 >;
 
 function ExplorerPlugin(props: GraphiQLExplorerPluginProps) {
@@ -152,9 +153,9 @@ function ExplorerPlugin(props: GraphiQLExplorerPluginProps) {
       checkboxUnchecked={checkboxUnchecked}
       checkboxChecked={checkboxChecked}
       styles={styles}
+      {...props}
       query={operationsString}
       onEdit={handleEditOperations}
-      {...props}
     />
   );
 }
